Separate Binaries: Move a Lot of Files
