+++ Cargo.toml
@@ -8,12 +8,8 @@ repository = "https://github.com/killercup/cargo-edit"
 version = "0.1.0"
 
 [[bin]]
-name = "main"
-path = "src/main.rs"
-
-[[bin]]
 name = "cargo-list"
-path = "src/bin/list.rs"
+path = "src/bin/list/main.rs"
 
 [features]
 default = []
