--- src/utils.rs
@@ -1,9 +1,0 @@
-macro_rules! toml_table {
-    ($key:expr => $value:expr) => {
-        {
-            let mut dep = BTreeMap::new();
-            dep.insert(String::from($key), toml::Value::String($value.clone()));
-            toml::Value::Table(dep)
-        }
-    }
-}
